+++ DelvUI/changelog.md
@@ -1,3 +1,7 @@
+# 2.3.1.6
+- Fixed Party Frames showing the wrong names.
+- Fixed Party Leader indicator.
+
 # 2.3.1.5
 - Fixed Party Frames status icons.